Update RN iOS setup to include RN 0.68 new Flipper initialization (#3905)
Summary:
When upgrading to RN 0.68, the upgrade helper removes the old Flipper initialization in favor of importing `RCTAppSetupUtils` which abstracts the Flipper initialization.
To be consistent, we should show in the docs that for 0.68 the preferred way is to import `RCTAppSetupUtils` in the AppDelegate.

## Changelog
- RN iOS setup docs has been updated to include the new way to initialize Flipper on React Native 0.68.

Pull Request resolved: https://github.com/facebook/flipper/pull/3905

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: flipper**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D37927365/V2/flipper/)|

|**Modified Pages**|
|[docs/getting-started/react-native-ios](https://our.intern.facebook.com/intern/staticdocs/eph/D37927365/V2/flipper/docs/getting-started/react-native-ios/)|

Reviewed By: lblasa, cortinico

Differential Revision: D37927365

Pulled By: passy

fbshipit-source-id: c6799e91a217d341eb817abb175ce63ea70c9597