Consider local visible rect and translation when calcauating view bounds
Summary: In general we want a child to know its position without having be overriden by it parent via offset child. Therefore we need to consider translation and the result of getLocalVisibleRect. Despite its name it seems to be an offset from its normal bounds after scrolling in certain types of scroll views.

Reviewed By: lblasa

Differential Revision: D40021836

fbshipit-source-id: 4ae9a4b202d7ba201cdb8b6bc8f13e94baf2b37e