Fix inability to create issue upon Publish failure (#4209)
Summary:
The publish job in the release workflow has been failing, but an issue has failed to be created due to the template file not being found. As per the error message, the template file isn't found because we need to check out the repo first.

## Changelog

Fix problem with auto-creating issue upon publish failure.

Pull Request resolved: https://github.com/facebook/flipper/pull/4209

Test Plan: Not able to run this locally, so we will need to wait upon the next release to ensure the issue gets created.

Reviewed By: mweststrate

Differential Revision: D40341329

Pulled By: aigoncharov

fbshipit-source-id: a9ad0a52a41ca13757fdd674937a075a024f618f