Update action in release workflow (#4220)
Summary:
The release workflow is using a version of the github release action that is broken, due to ghr now being dynamically linked and not compatible with Alpine Linux's minimal version of libc. The fix is to use the version of the action which installs gcompat so that ghr can run on the docker image.

## Changelog

Update release workflow to use version of action compatible with latest version of ghr.

Pull Request resolved: https://github.com/facebook/flipper/pull/4220

Test Plan: N/A

Reviewed By: nikoant

Differential Revision: D40443738

Pulled By: aigoncharov

fbshipit-source-id: 7765420fbe034aa80ce07c698ff83773c81ed52f