Fix bounds for children of nested scroll view
Summary:
Android does not set the top/left of the child of view pager. I attempted to solve this by grabbing the offset from the scrollview and passing it down via offset child. The issue is offsetchild has a lot of problems with the partial layout traversal. If the child of the scroll view is an observer root then the layout traversal with short circuit and setup a new observer for the child but we lose the offset child in the process.

This solution is a little dirty but it is functional for now

Reviewed By: lblasa

Differential Revision: D40430778

fbshipit-source-id: 7ad53b2a4818b55515e7662272548e2ce17e6b44