Fragment return 0 bounds
Summary: In a previous diff the Bounds was made mandatory and the fragment bounds was set to be the same as the underlying view. this caused the offset to be doubled since the underlying view would also apply that offset. Since fragments dont contribute to layout and are purely for lifecycle we can set it to 0 and handle it on the desktop

Reviewed By: lblasa

Differential Revision: D41304500

fbshipit-source-id: 39591b3ef0c746bd7a4ec8d61fed623cc7c94944