Fixing issue where certain activities not tracked
Summary:
The root view resolver will always find all root views but there was a bug in the listrootviews method. The code was very complex and most of the code seemed to be unneeded. Its now working. The listrootviews method now just returns teh contents of the observable array.

The reason we needed this is that Certain activities dont seem to tracked by the listener we add to `registerActivityLifecycleCallbacks` Its as if there is a floating decor with no activity. One example in FB4a is clicking on a notification in the notifications panel

Reviewed By: lblasa

Differential Revision: D41522791

fbshipit-source-id: b49b0104ddf758f097e1fd3f9ac6588de2d3646e