Android activity tracking uses root view resolver as sole source of truth
Summary:
The previous approach was to append any 'extra/floating' root view to the end of the childrens for the application descriptor. But the active child was always the last view so if there was a 'floating' decor view with no activity it would always be the active child, even if you opened a real activity from it. This was possible to do in FB4a and you could get into a situation where you couldn't see the top most activity.

New approach is to only iterate the root views from the resolver since it contains everything and map to activity if possible.

Reviewed By: lblasa

Differential Revision: D41731951

fbshipit-source-id: 9ef3fcfdaa41e9b07606ca6781cf22482b54e072