Fix style guide internally
Summary:
The site loads the style guide in an iframe so that it's css is isolated from the main sites styles.

But it doesn't know how big to render the iframe, because it doesn't know the size of its contents.
Previously, the js running inside the iframe measured itself and then accessed itself using the parents frame and modified it's own size to match.

That doesn't work anymore because the iframe isn't allowed to interact directly with its parent.
So instead, i'm using message passing.

Also changing MutationObserver to ResizeObserver because the former does not always fire when new css gets loaded and applied.

Reviewed By: aigoncharov

Differential Revision: D41735809

fbshipit-source-id: 9e84a9b20741f5470012e25f240fb6a2b494f7e3