Add support for non support libarry scroll views
Summary: we had special handling previously for nested scrollview and we previously used the indirect approach of asking for the local visible rect. New approach is to get the scroll x from the parent and apply it, will work for all scrolling view and it is a lot clearer what the intention is.

Reviewed By: lblasa

Differential Revision: D41772106

fbshipit-source-id: 04b5e68ecabd70548e788ed18423a360ce6c3fa7