Plugin RPC errors should not cause crash report
Summary:
This code path is when a send results in an error being returned from the client.  We reject the promise but also raise an error on the message bus.

There is a handler for this 'error' event that is raised that is over zelous and reports a crash to the crash reporter. This should not happen

Reviewed By: lblasa

Differential Revision: D42385292

fbshipit-source-id: f668a396b0d266ee9d1c7c1ca740e7bb5ae9608a